scottstock {scottstock}	R Documentation

Yahoo stock data wrapper of BatchGetSymbols for Finance class

Description

Pulls stock data from Yahoo Finance and then do analysis required for Finance II class

Usage

scottstock(ticker, frequency, start_date, end_date)

Arguments

ticker
The stock symbol you want: "MSFT", or you can specify multiple by c("MSFT", "FB")

frequency
The type of data you want to pull: "daily", "weekly", "monthly", "yearly"

start_date
The earliest date from which you want to pull data: "MM-DD-YYYY" surrounded by quotes

end_date
The latest date from which you want to pull data: "MM-DD-YYYY" surrounded by quotes

References

This script is a wrapper for Marcelo Perlin's BatchGetSymbols package that also automates analysis and regression commands.

Examples

## Example of function usage:
myStock <- scottstock("MSFT","monthly",
    "01-01-2019", "12-31-2019")

[Package scottstock version 1.3.0 Index]